requirements.txt
Pillow
yt-dlp
